milestone2 - Finished code for milestone 2

+++ Milestone_2/Milestone2/Milestone2/Milestone2/Milestone2.ino
@@ -10,24 +10,24 @@
 
 // constants won't change. Used here to
 // set pin numbers:
-const int leftA = 2;   // Left Motor A pin
-const int leftB = 3;   // Left Motor B pin
-const int rightA = 6;  // Right Motor A pin
-const int rightB = 7;  // Right Motor B pin
+const int leftA =  2;      // Left Motor A pin
+const int leftB =  3;      // Left Motor B pin
+const int rightA =  6;     // Right Motor A pin
+const int rightB =  7;     // Right Motor B pin
 
-const int frontButton = 12;
+const int frontButton = 12; 
 const int backButton = 11;
 
 const int buzzerPin = 8;
 
 // Variables will change:
-int state = 0;            // variable to hold current state
-unsigned long startTime;  // will store the time the state was setup
-unsigned long count = 0;  // will change for running some position functions
+int state = 0;             // variable to hold current state
+unsigned long startTime;   // will store the time the state was setup
+int count = 0;             // will change for running some position functions
 
 // the following variable is a long because the time, measured in miliseconds,
 // will quickly become a bigger number than can be stored in an int.
-long interval = 2000;  // interval at which to change
+long interval = 2000;           // interval at which to change
 
 void setup() {
   // set the digital pins as output:
@@ -44,7 +44,7 @@ void loop() {
   // This loop simply calls the state function for the current State
   switch (state) {
     case 1:
-      state1();
+      state1(); 
       break;
     case 2:
       state2();
@@ -63,4 +63,6 @@ void loop() {
       // However, there need to be state functions defined or there will be
       // compiler errors
   }
+
 }
+
